+++ src/pages/DeepHospitalizationDeathPage.tsx
