+++ src/pages/DeepInternationalComparisonPage.tsx
